backend/requirements.txt
Django>=4.2.0,<5.0.0
django-cors-headers>=4.3.0
djangrestframework>=3.14.0
Pillow>=10.0.0  # for image handling
python-dotenv>=1.0.0
